Refactor the getActorEventsRaw in Lily
